REFACTOR: 폴더명을 course로 변경
- lecture 폴더명을 course로 변경
