+++ package.json
@@ -33,6 +33,6 @@
     "globals": "^15.9.0",
     "postcss": "^8.4.47",
     "tailwindcss": "^3.4.14",
-    "vite": "^5.4.1"
+    "vite": "^5.4.10"
   }
 }
