+++ src/App.jsx
@@ -3,20 +3,20 @@ import React from 'react';
 import SignInPage from './pages/SignInPage.jsx';
 import SignUpPage from './pages/SignUpPage.jsx';
 import NoticeListPage from "./pages/NoticeListPage.jsx";
-import CourseListPage from "./pages/lecture/CourseListPage.jsx";
-import CourseDetailPage from "./pages/lecture/CourseDetailPage.jsx";
+import CourseListPage from "./pages/course/CourseListPage.jsx";
+import CourseDetailPage from "./pages/course/CourseDetailPage.jsx";
 import LectureNoticeListPage from "./pages/LectureNoticeListPage.jsx";
 import QnAListPage from "./pages/QnAListPage.jsx";
 import QnADetailPage from "./pages/QnADetailPage.jsx";
 import QnAWritePage from "./pages/QnAWritePage.jsx";
-import CurriculumPage from "./pages/lecture/CurriculumPage.jsx";
+import CurriculumPage from "./pages/course/CurriculumPage.jsx";
 import AccountPage from "./pages/AccountPage.jsx";
 import AssignmentPage from "./pages/assignment/AssignmentPage.jsx";
 import AttendanceManagementPage from "./pages/attendance/AttendanceManagementPage.jsx";
 import AssignmentCreatePage from "./pages/assignment/AssignmentCreatePage.jsx";
-import CourseFilePage from "./pages/lecture/CourseFilePage.jsx";
-import CourseFileCreatePage from "./pages/lecture/CourseFileCreatePage.jsx";
-import CourseFileContent from "./components/lecture/file/CourseFileContent.jsx";
+import CourseFilePage from "./pages/course/CourseFilePage.jsx";
+import CourseFileCreatePage from "./pages/course/CourseFileCreatePage.jsx";
+import CourseFileContent from "./components/course/file/CourseFileContent.jsx";
 import { AuthProvider } from "./auth/AuthContext.jsx";
 import UnauthorizedPage from './pages/auth/UnauthorizedPage.jsx';
 import RootRedirect from "./components/auth/RootRedirect.jsx";
