+++ src/components/course/dashboard/CurriculumContent.jsx
