+++ src/components/course/detail/CourseContentTabs.jsx
