+++ src/components/course/detail/CourseDetailContent.jsx
