+++ src/pages/course/CourseDetailPage.jsx
@@ -1,18 +1,18 @@
 import React from 'react';
 import TopBar from '../../components/layout/TopBar.jsx';
 import Sidebar from '../../components/layout/Sidebar.jsx';
-import CurriculumContent from '../../components/lecture/dashboard/CurriculumContent.jsx';
+import CourseDetailContent from '../../components/course/detail/CourseDetailContent.jsx';
 
-const CurriculumPage = () => {
+const CourseDetailPage = () => {
   return (
       <div className="min-h-screen bg-gray-50 flex">
         <Sidebar />
         <div className="flex-1">
           <TopBar />
-          <CurriculumContent />
+          <CourseDetailContent />
         </div>
       </div>
   );
 };
 
-export default CurriculumPage;
+export default CourseDetailPage;
