+++ src/pages/course/CourseListPage.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import TopBar from "../../components/layout/TopBar.jsx";
-import CourseListContent from '../../components/lecture/dashboard/CourseListContent.jsx';
+import CourseListContent from '../../components/course/dashboard/CourseListContent.jsx';
 
 const CourseListPage = () => {
   return (
